feat: 改进事件模型
